add server-side validations for notes

+++ src/main/java/me/remil/notes/service/notes/NotesServiceImpl.java
@@ -4,6 +4,7 @@ import me.remil.notes.dao.NotesRepository;
 import me.remil.notes.dto.send.NoteTitles;
 import me.remil.notes.entity.Notes;
 import me.remil.notes.exception.IdAlreadyExistsException;
+import me.remil.notes.exception.MissingItemException;
 import me.remil.notes.exception.NotFoundException;
 import me.remil.notes.exception.UnauthorizedRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -39,34 +41,82 @@ public class NotesServiceImpl implements NotesService {
     }
 
     @Override
-    public Notes fetchNoteById(String noteId) {
+    public Notes fetchNoteById(String noteId, String usernameFromToken) {
+        String username = notesRepository.fetchUserIdByNoteId(noteId);
+        if (!username.equals(usernameFromToken)) {
+            throw new UnauthorizedRequestException("UNAUTHORIZED: Request cannot be fulfilled");
+        }
         return notesRepository.findByNoteId(noteId);
     }
 
     @Override
     public void saveNote(Notes note) {
-        if (notesRepository.countByNoteId(note.getNoteId()) != 0) {
-            throw new IdAlreadyExistsException("Note id already exists");
-        }
         notesRepository.save(note);
     }
 
     @Override
-    public void updateNote(Notes note, String userId) {
+    public void updateNote(Notes note, String usernameFromToken) {
         String userIdFromDb = notesRepository.fetchUserIdByNoteId(note.getNoteId());
 
         if (userIdFromDb == null) {
-            throw new NotFoundException("No note found with that Id");
+            throw new NotFoundException("No note associated with that id");
         }
 
-        if (!userId.equals(userIdFromDb)) {
+        if (!usernameFromToken.equals(userIdFromDb)) {
             throw new UnauthorizedRequestException("Requesting access to unauthorized resources.");
         }
         notesRepository.save(note);
     }
 
     @Override
-    public void deleteNote(String noteId) {
+    public void deleteNote(String noteId, String usernameFromToken) {
+        String username = notesRepository.fetchUserIdByNoteId(noteId);
+        if (!username.equals(usernameFromToken)) {
+            throw new UnauthorizedRequestException("UNAUTHORIZED: Request cannot be fulfilled");
+        }
         notesRepository.deleteById(noteId);
+    }
+
+    @Override
+    public void validateBeforeSaveOrUpdate(Notes note, String usernameFromToken, ACTIONS type) {
+        String noteId = note.getNoteId();
+        String noteTitle = note.getNoteHeading();
+        String username = note.getUserId();
+        String noteBody = note.getNoteBody();
+        Timestamp lastUpdated = note.getLastUpdated();
+
+        if (username == null) {
+            throw new MissingItemException("Missing user credentials in request");
+        } else if (!username.equals(usernameFromToken)) {
+            throw new UnauthorizedRequestException("UNAUTHORIZED: Request cannot be fulfilled");
+        }
+
+        if (noteId == null || noteId.trim().length() == 0) {
+            throw new MissingItemException("Note id is missing");
+        }
+
+        if (noteTitle == null || noteTitle.trim().length() == 0) {
+            throw new MissingItemException("Missing title for note");
+        }
+
+        if (noteBody == null) {
+            throw new MissingItemException("No note given");
+        }
+
+        if (lastUpdated == null) {
+            throw new MissingItemException("Missing last updated time");
+        }
+
+        boolean noteExists = notesRepository.existsByNoteId(noteId);
+
+        if (type.equals(ACTIONS.SAVE)) {
+            if (noteExists) {
+                throw new IdAlreadyExistsException("Note with id "+noteId+" already exists");
+            }
+            saveNote(note);
+        } else if (!noteExists) {
+            throw new NotFoundException("No note found with id "+noteId);
+        }
+        updateNote(note, usernameFromToken);
     }
 }
